// Capstone Project/src/app/pages/admin/admin-room-type-management-page/screen/types.ts
export enum adminPageRoomTypeEnum {
  GET_DATA_ROOM_TYPE_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_ROOM_TYPE_REQUEST',
  GET_DATA_ROOM_TYPE_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_ROOM_TYPE_SUCCESS',
  GET_DATA_ROOM_TYPE_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_ROOM_TYPE_FAILURE',

  CREATE_DATA_ROOM_TYPE_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_ROOM_TYPE_REQUEST',
  CREATE_DATA_ROOM_TYPE_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_ROOM_TYPE_SUCCESS',
  CREATE_DATA_ROOM_TYPE_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_ROOM_TYPE_FAILURE',

  UPDATE_DATA_ROOM_TYPE_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_ROOM_TYPE_REQUEST',
  UPDATE_DATA_ROOM_TYPE_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_ROOM_TYPE_SUCCESS',
  UPDATE_DATA_ROOM_TYPE_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_ROOM_TYPE_FAILURE',

  DELETE_DATA_ROOM_TYPE_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_ROOM_TYPE_REQUEST',
  DELETE_DATA_ROOM_TYPE_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_ROOM_TYPE_SUCCESS',
  DELETE_DATA_ROOM_TYPE_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_ROOM_TYPE_FAILURE',

  CLEAR_ROOM_TYPE_STATE = 'APP/ADMIN_PAGE/CLEAR_ROOM_TYPE_STATE',
}